Make V2 of aquapath obsolete and implement changes to V1
