+++ electron/src/machines/properties.ts
@@ -488,23 +488,12 @@ export const aquapath1: MachineProperties = {
     },
     {
       role: 3,
-      role_label: "EL3204",
+      role_label: "EL3024",
       allowed_devices: [
         {
           vendor_id: 2,
-          product_id: 0xc843052,
-          revision: 0x160000,
-        },
-      ],
-    },
-    {
-      role: 4,
-      role_label: "EL5152",
-      allowed_devices: [
-        {
-          vendor_id: 2,
-          product_id: 0x14203052,
-          revision: 0x140000,
+          product_id: 0xbd03052,
+          revision: 0x130000,
         },
       ],
     },
@@ -947,84 +936,6 @@ export const bottlecapsTestMachine: MachineProperties = {
   ],
 };
 
-export const aquapath2: MachineProperties = {
-  name: "Aquapath",
-  version: "V2",
-  slug: "aquapath2",
-  icon: "lu:Waves",
-  machine_identification: {
-    vendor: VENDOR_QITECH,
-    machine: 0x000b,
-  },
-  device_roles: [
-    {
-      role: 0,
-      role_label: "Bus Coupler",
-      allowed_devices: [
-        {
-          vendor_id: 2,
-          product_id: 0x44c2c52,
-          revision: 0x120000,
-        },
-      ],
-    },
-    {
-      role: 1,
-      role_label: "EL2008",
-      allowed_devices: [
-        {
-          vendor_id: 2,
-          product_id: 0x7d83052,
-          revision: 0x110000,
-        },
-        {
-          vendor_id: 2,
-          product_id: 0x7d83052,
-          revision: 0x120000,
-        },
-      ],
-    },
-    {
-      role: 2,
-      role_label: "EL3204",
-      allowed_devices: [
-        {
-          vendor_id: 2,
-          product_id: 0xc843052,
-          revision: 0x160000,
-        },
-        {
-          vendor_id: 2,
-          product_id: 0xc843052,
-          revision: 0x150000,
-        },
-      ],
-    },
-    {
-      role: 3,
-      role_label: "EL4002",
-      allowed_devices: [
-        {
-          vendor_id: 2,
-          product_id: 0xfa23052,
-          revision: 0x140000,
-        },
-      ],
-    },
-    {
-      role: 4,
-      role_label: "EL3024",
-      allowed_devices: [
-        {
-          vendor_id: 2,
-          product_id: 0xbd03052,
-          revision: 0x130000,
-        },
-      ],
-    },
-  ],
-};
-
 export const machineProperties: MachineProperties[] = [
   winder2,
   winder2_spool_7031,
@@ -1035,7 +946,6 @@ export const machineProperties: MachineProperties[] = [
   buffer1,
   wagoPower1,
   aquapath1,
-  aquapath2,
   testmachine,
   analogInputTestMachine,
   wagoAiTestMachine,
