--- machines/src/aquapath2/api.rs
@@ -1,326 +1,0 @@
-use super::{AquaPathV2, AquaPathV2Mode, controller::CoolingMode};
-use crate::{MachineApi, MachineMessage};
-use control_core::socketio::{
-    event::{Event, GenericEvent},
-    namespace::{
-        CacheFn, CacheableEvents, Namespace, NamespaceCacheingLogic, cache_first_and_last_event,
-    },
-};
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use smol::channel::Sender;
-use std::sync::Arc;
-use tracing::instrument;
-
-#[derive(Serialize, Debug, Clone, Default)]
-pub struct LiveValuesEvent {
-    pub front_flow: f64,
-    pub back_flow: f64,
-    pub front_temperature: f64,
-    pub back_temperature: f64,
-    pub front_temp_reservoir: f64,
-    pub back_temp_reservoir: f64,
-    pub front_revolutions: f64,
-    pub back_revolutions: f64,
-    pub front_power: f64,
-    pub back_power: f64,
-    pub front_heating: bool,
-    pub back_heating: bool,
-    pub front_cooling_mode: Option<CoolingMode>,
-    pub back_cooling_mode: Option<CoolingMode>,
-    pub front_pump_cooldown_active: bool,
-    pub back_pump_cooldown_active: bool,
-    pub front_pump_cooldown_remaining: f64,
-    pub back_pump_cooldown_remaining: f64,
-    pub front_heating_startup_wait_active: bool,
-    pub back_heating_startup_wait_active: bool,
-    pub front_heating_startup_wait_remaining: f64,
-    pub back_heating_startup_wait_remaining: f64,
-    pub front_total_energy: f64,
-    pub back_total_energy: f64,
-}
-
-impl LiveValuesEvent {
-    pub fn build(&self) -> Event<Self> {
-        Event::new("LiveValuesEvent", self.clone())
-    }
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct StateEvent {
-    pub is_default_state: bool,
-    pub mode_state: ModeState,
-    pub ambient_temperature_calibration: f64,
-    pub default_heating_tolerance: f64,
-    pub default_cooling_tolerance: f64,
-    pub default_pid_kp: f64,
-    pub default_pid_ki: f64,
-    pub default_pid_kd: f64,
-    pub flow_states: FlowStates,
-    pub temperature_states: TempStates,
-    pub fan_states: FanStates,
-    pub cooling_mode_states: CoolingModeStates,
-    pub tolerance_states: ToleranceStates,
-    pub pid_states: PidStates,
-    pub thermal_safety_states: ThermalSafetyStates,
-}
-
-impl StateEvent {
-    pub fn build(&self) -> Event<Self> {
-        Event::new("StateEvent", self.clone())
-    }
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct NoticeEvent {
-    pub title: String,
-    pub message: String,
-}
-
-impl NoticeEvent {
-    pub fn build(&self) -> Event<Self> {
-        Event::new("NoticeEvent", self.clone())
-    }
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct TempStates {
-    pub front: TempState,
-    pub back: TempState,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct TempState {
-    pub temperature: f64,
-    pub target_temperature: f64,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct ModeState {
-    pub mode: AquaPathV2Mode,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct FlowStates {
-    pub front: FlowState,
-    pub back: FlowState,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct FlowState {
-    pub flow: f64,
-    pub should_flow: bool,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct FanState {
-    pub revolutions: f64,
-    pub max_revolutions: f64,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct FanStates {
-    pub front: FanState,
-    pub back: FanState,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct CoolingModeState {
-    pub mode: Option<CoolingMode>,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct CoolingModeStates {
-    pub front: CoolingModeState,
-    pub back: CoolingModeState,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct ToleranceState {
-    pub heating: f64,
-    pub cooling: f64,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct ToleranceStates {
-    pub front: ToleranceState,
-    pub back: ToleranceState,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct PidState {
-    pub kp: f64,
-    pub ki: f64,
-    pub kd: f64,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct PidStates {
-    pub front: PidState,
-    pub back: PidState,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct ThermalSafetyState {
-    pub thermal_delay: f64,
-    pub cooldown_min_temperature: f64,
-}
-
-#[derive(Serialize, Debug, Clone)]
-pub struct ThermalSafetyStates {
-    pub front: ThermalSafetyState,
-    pub back: ThermalSafetyState,
-}
-
-pub enum AquaPathV2Events {
-    LiveValues(Event<LiveValuesEvent>),
-    State(Event<StateEvent>),
-    Notice(Event<NoticeEvent>),
-}
-
-#[derive(Deserialize, Serialize)]
-enum Mutation {
-    SetAquaPathMode(AquaPathV2Mode),
-
-    SetFrontTemperature(f64),
-    SetBackTemperature(f64),
-
-    SetFrontFlow(bool),
-    SetBackFlow(bool),
-
-    SetFrontRevolutions(f64),
-    SetBackRevolutions(f64),
-
-    SetFrontHeatingTolerance(f64),
-    SetBackHeatingTolerance(f64),
-    SetFrontCoolingTolerance(f64),
-    SetBackCoolingTolerance(f64),
-    SetFrontPidKp(f64),
-    SetFrontPidKi(f64),
-    SetFrontPidKd(f64),
-    SetBackPidKp(f64),
-    SetBackPidKi(f64),
-    SetBackPidKd(f64),
-    SetFrontThermalFlowSettleDuration(f64),
-    SetBackThermalFlowSettleDuration(f64),
-    SetFrontPumpCooldownMinTemperature(f64),
-    SetBackPumpCooldownMinTemperature(f64),
-    SetAmbientTemperatureCalibration(f64),
-}
-
-#[derive(Debug, Clone)]
-pub struct AquaPathV2Namespace {
-    pub namespace: Option<Namespace>,
-}
-
-impl NamespaceCacheingLogic<AquaPathV2Events> for AquaPathV2Namespace {
-    #[instrument(skip_all)]
-    fn emit(&mut self, events: AquaPathV2Events) {
-        let event = Arc::new(events.event_value());
-        let buffer_fn = events.event_cache_fn();
-        match &mut self.namespace {
-            Some(ns) => ns.emit(event, &buffer_fn),
-            None => (),
-        }
-    }
-}
-
-impl CacheableEvents<AquaPathV2Events> for AquaPathV2Events {
-    fn event_value(&self) -> GenericEvent {
-        match self {
-            AquaPathV2Events::LiveValues(event) => event.into(),
-            AquaPathV2Events::State(event) => event.into(),
-            AquaPathV2Events::Notice(event) => event.into(),
-        }
-    }
-
-    fn event_cache_fn(&self) -> CacheFn {
-        let cache_first_and_last = cache_first_and_last_event();
-        match self {
-            AquaPathV2Events::LiveValues(_) => cache_first_and_last,
-            AquaPathV2Events::State(_) => cache_first_and_last,
-            AquaPathV2Events::Notice(_) => Box::new(|_, _| {}),
-        }
-    }
-}
-
-impl MachineApi for AquaPathV2 {
-    fn api_get_sender(&self) -> Sender<MachineMessage> {
-        self.api_sender.clone()
-    }
-
-    fn api_mutate(&mut self, request_body: Value) -> Result<(), anyhow::Error> {
-        let control: Mutation = serde_json::from_value(request_body)?;
-        match control {
-            Mutation::SetAquaPathMode(mode) => self.set_mode_state(mode),
-            Mutation::SetBackTemperature(temperature) => {
-                self.set_target_temperature(temperature, super::AquaPathSideType::Back)
-            }
-            Mutation::SetFrontTemperature(temperature) => {
-                self.set_target_temperature(temperature, super::AquaPathSideType::Front)
-            }
-            Mutation::SetBackFlow(should_pump) => {
-                self.set_should_pump(should_pump, super::AquaPathSideType::Back)
-            }
-            Mutation::SetFrontFlow(should_pump) => {
-                self.set_should_pump(should_pump, super::AquaPathSideType::Front)
-            }
-            Mutation::SetBackRevolutions(revolution) => {
-                self.set_max_revolutions(revolution, super::AquaPathSideType::Back)
-            }
-            Mutation::SetFrontRevolutions(revolutions) => {
-                self.set_max_revolutions(revolutions, super::AquaPathSideType::Front)
-            }
-            Mutation::SetBackHeatingTolerance(tolerance) => {
-                self.set_heating_tolerance(tolerance, super::AquaPathSideType::Back)
-            }
-            Mutation::SetFrontHeatingTolerance(tolerance) => {
-                self.set_heating_tolerance(tolerance, super::AquaPathSideType::Front)
-            }
-            Mutation::SetBackCoolingTolerance(tolerance) => {
-                self.set_cooling_tolerance(tolerance, super::AquaPathSideType::Back);
-            }
-            Mutation::SetFrontCoolingTolerance(tolerance) => {
-                self.set_cooling_tolerance(tolerance, super::AquaPathSideType::Front);
-            }
-            Mutation::SetFrontPidKp(value) => {
-                self.set_pid_kp(value, super::AquaPathSideType::Front);
-            }
-            Mutation::SetFrontPidKi(value) => {
-                self.set_pid_ki(value, super::AquaPathSideType::Front);
-            }
-            Mutation::SetFrontPidKd(value) => {
-                self.set_pid_kd(value, super::AquaPathSideType::Front);
-            }
-            Mutation::SetBackPidKp(value) => {
-                self.set_pid_kp(value, super::AquaPathSideType::Back);
-            }
-            Mutation::SetBackPidKi(value) => {
-                self.set_pid_ki(value, super::AquaPathSideType::Back);
-            }
-            Mutation::SetBackPidKd(value) => {
-                self.set_pid_kd(value, super::AquaPathSideType::Back);
-            }
-            Mutation::SetFrontThermalFlowSettleDuration(value) => {
-                self.set_thermal_flow_settle_duration(value, super::AquaPathSideType::Front);
-            }
-            Mutation::SetBackThermalFlowSettleDuration(value) => {
-                self.set_thermal_flow_settle_duration(value, super::AquaPathSideType::Back);
-            }
-            Mutation::SetFrontPumpCooldownMinTemperature(value) => {
-                self.set_pump_cooldown_min_temperature(value, super::AquaPathSideType::Front);
-            }
-            Mutation::SetBackPumpCooldownMinTemperature(value) => {
-                self.set_pump_cooldown_min_temperature(value, super::AquaPathSideType::Back);
-            }
-            Mutation::SetAmbientTemperatureCalibration(ambient_temp) => {
-                self.set_ambient_temperature_calibration(ambient_temp);
-            }
-        }
-        Ok(())
-    }
-
-    fn api_event_namespace(&mut self) -> Option<Namespace> {
-        self.namespace.namespace.clone()
-    }
-}
